Rename package from ap_move_lights to ap_move_raw_light_to_blink
- Rename package directory to match repository name
- Update pyproject.toml package name and entry points
- Update Makefile targets to use new package name
- Update MANIFEST.in for new package directory
- Update README.md with new package name and install URLs
- Update all Python imports and test docstrings
