+++ MANIFEST.in
@@ -2,4 +2,4 @@ include README.md
 include LICENSE
 include Makefile
 include pyproject.toml
-recursive-include ap_move_lights *.py
+recursive-include ap_move_raw_light_to_blink *.py
