+++ ap_move_raw_light_to_blink/__main__.py
@@ -1,7 +1,7 @@
 # Generated By: Claude Code (Claude Sonnet 4.5)
 """Entry point for python -m invocation."""
 
-from ap_move_lights.move_lights import main
+from ap_move_raw_light_to_blink.move_lights import main
 
 if __name__ == "__main__":
     main()
