+++ tests/test_move_lights.py
@@ -1,5 +1,5 @@
 """
-Unit tests for ap_move_lights.move_lights module.
+Unit tests for ap_move_raw_light_to_blink.move_lights module.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
@@ -7,7 +7,7 @@ Generated By: Cursor (Claude Sonnet 4.5)
 import os
 from unittest.mock import patch
 
-from ap_move_lights import move_lights, config
+from ap_move_raw_light_to_blink import move_lights, config
 
 
 class TestMoveFiles:
@@ -283,7 +283,7 @@ class TestMoveFiles:
 class TestMain:
     """Tests for main function."""
 
-    @patch("ap_move_lights.move_lights.move_files")
+    @patch("ap_move_raw_light_to_blink.move_lights.move_files")
     def test_main_calls_move_files(self, mock_move_files):
         """Test that main function calls move_files with correct arguments."""
         with patch("sys.argv", ["move_lights.py", "/source", "/dest"]):
@@ -299,7 +299,7 @@ class TestMain:
             create_accept=True,
         )
 
-    @patch("ap_move_lights.move_lights.move_files")
+    @patch("ap_move_raw_light_to_blink.move_lights.move_files")
     def test_main_with_debug_flag(self, mock_move_files):
         """Test main with --debug flag."""
         with patch("sys.argv", ["move_lights.py", "/source", "/dest", "--debug"]):
@@ -315,7 +315,7 @@ class TestMain:
             create_accept=True,
         )
 
-    @patch("ap_move_lights.move_lights.move_files")
+    @patch("ap_move_raw_light_to_blink.move_lights.move_files")
     def test_main_with_dryrun_flag(self, mock_move_files):
         """Test main with --dryrun flag."""
         with patch("sys.argv", ["move_lights.py", "/source", "/dest", "--dryrun"]):
@@ -331,7 +331,7 @@ class TestMain:
             create_accept=True,
         )
 
-    @patch("ap_move_lights.move_lights.move_files")
+    @patch("ap_move_raw_light_to_blink.move_lights.move_files")
     def test_main_with_both_flags(self, mock_move_files):
         """Test main with both --debug and --dryrun flags."""
         with patch(
